ACT/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
anytree==2.12.1
appnope==0.1.4
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
blinker==1.7.0
blis==0.7.11
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.1
commonmark==0.9.1
confection==0.1.4
contourpy==1.2.0
coverage==7.4.4
cryptography==42.0.5
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.49.0
fsspec==2024.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grobid-client-python==0.0.8
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
identify==2.5.35
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.3
iniconfig==2.0.0
ipykernel==6.29.3
ipympl==0.9.4
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
keybert==0.8.4
kiwisolver==1.4.5
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
openai==1.14.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
parso==0.8.3
pdfalyzer==1.14.6
pdfminer.six==20231228
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.2
PyPDF2==2.12.1
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.3
Redis-Sentinel-Url==1.0.1
regex==2023.12.25
requests==2.31.0
rich==12.6.0
rich_argparse_plus==0.3.1.4
rq==1.16.1
rq-dashboard==0.6.7.2
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
thinc==8.2.3
threadpoolctl==3.3.0
tika==2.6.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typer==0.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
utils==1.0.2
virtualenv==20.25.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
yara-python==4.2.3
yaralyzer==0.9.3
zipp==3.17.0
